% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.r
\name{check_missing}
\alias{check_missing}
\title{Check and summarize missing variables}
\usage{
check_missing(data)
}
\arguments{
\item{data}{data.frame obj}
}
\description{
Steps and functions for checking and summarizing missing variables
}
\details{
The output will be a summary of missing variables
}
