% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_select.r
\name{rm_near_zero_vars}
\alias{rm_near_zero_vars}
\alias{rm_high_correlated}
\alias{rm_linear_combos}
\title{Remove non informative variables}
\usage{
rm_near_zero_vars(data, keep = NULL, freqCut = 95/5, uniqueCut = 10,
  verbose = FALSE)

rm_high_correlated(data, keep = NULL, cutoff = 0.9, verbose = FALSE)

rm_linear_combos(data, keep = NULL, verbose = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{keep}{vector of variables to keep}

\item{freqCut}{first/second, default 95/5}

\item{uniqueCut}{nfac/nrow, default 10(\%)}

\item{verbose}{default FALSE}

\item{cutoff}{default 0.9}
}
\description{
This is wrap of `nearZeroVar()`,`findCorrelation()` from caret
}
\details{
The output of all these steps will be a dataframe as original class
}
