% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_continuous_corrplot.r
\name{ggplot_continuous_corrplot}
\alias{ggplot_continuous_corrplot}
\title{Correlation plot of continuous vars}
\usage{
ggplot_continuous_corrplot(data, vars, tar, threshold = 0.5)
}
\arguments{
\item{data}{dataframe}

\item{vars}{character vector of vars}

\item{tar}{character of target variable name}

\item{threshold}{default 0.5}
}
\description{
This is a wrap of correlation matrix plot code
}
\details{
The output of `ggplot_continuous_corrplot()` will be a correlation matrix plot
}
