% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_vars_eda.r
\name{ggplot_continuous_hist}
\alias{ggplot_continuous_hist}
\alias{ggplot_factor_barplot}
\alias{ggplot_factor_continuous_boxplot}
\alias{ggplot_factor_continuous_hist}
\alias{ggplot_continuous_continuous_scatter}
\title{Data exploration plotting family based on ggplot}
\usage{
ggplot_continuous_hist(data, x, facet = NULL, binwidth = "auto",
  ggtitle = sprintf("Histogram of \%s", x), xlab = sprintf("\%s", x),
  ylab = sprintf("Count of \%s", x))

ggplot_factor_barplot(data, x, facet = NULL,
  ggtitle = sprintf("Count of \%s", x), ...)

ggplot_factor_continuous_boxplot(data, x, y, facet = NULL,
  ggtitle = sprintf("Boxplot of \%s by \%s", y, x))

ggplot_factor_continuous_hist(data, x, y, facet = NULL, binwidth = "auto",
  ggtitle = sprintf("Histogram of \%s by \%s", y, x))

ggplot_continuous_continuous_scatter(data, x, y, facet = NULL,
  ggtitle = sprintf("Scatter plot of \%s and \%s", x, y))
}
\arguments{
\item{data}{dataframe objects}

\item{x}{string name in data, project as x-axis}

\item{facet}{vector of string as facet in ggplot}

\item{binwidth}{default 5000}

\item{ggtitle}{graph title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{y}{string name in data, project as y-axis}
}
\description{
A interface for plotting on data exploration
}
\details{
This output will be a ggplot style plot
}
\examples{
ggplot_continuous_hist(data,"ACV_1612",c("tag","CHAIN2"))
}
