% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_n_require.r
\name{install_n_require}
\alias{install_n_require}
\title{Install package if not required}
\usage{
install_n_require(pkglist)
}
\arguments{
\item{pkglist}{vector of package names in character that need to required for installation}
}
\value{
`install_n_require()` returns an error message when error
}
\description{
This is a basic interface of installing required package automatically when need
The output of `install_n_require()` will be a error of correponding error when installation failed, otherwise quiet
}
\examples{
install_n_require(pkglist = c('ggplot2','ggfortify'))

}
