% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.r
\name{multiplot.auto}
\alias{multiplot.auto}
\alias{multiplot.manual}
\title{multiple ggplot in one interface}
\usage{
multiplot.auto(data, xvars, yvars = NULL, FUN = ggplot_continuous_hist)

multiplot.manual(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{xvars}{x axis vars}

\item{yvars}{x axis vars}

\item{cols}{number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}

\item{.../plotlist}{ggplot obj list separated by comma,
or bracket use `list()` function}
}
\description{
Try to place multiple ggplot in one interface
}
\details{
This output will be ggplot style multiple plot.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right,
and 3 will go all the way across the bottom.
}
\examples{
multiplot.auto(data = Glass,x = colnames(Glass)[1: 9],y = colnames(Glass)[9],FUN = ggplot_continuous_continuous_scatter)
}
